+++ extension/Changelog.md
@@ -5,6 +5,7 @@
 - Change color of settings text on hover to better show it being able to be clicked
 - Added API (feel free to request access to additional functions)
 - Support male genitals
+- Possibility for automatic erections based on arousal
 
 ### Bugfixes:
 - Vibrators on zones that cant make you orgasm, still gave orgasm progress